rename package names as detail and search
