+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/search/RepositorySearchFragment.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check.reposearch
+package jp.co.yumemi.android.code_check.search
 
 import android.os.Bundle
 import android.view.LayoutInflater
